Find Rust's source path
This is scraped from rustic-mode / rust-mode's function to do exactly the
same thing, but adds lib/rustlib/src/rust/library to the search list because
that seems to be where the stdlib is located on at least one of my machines.

I'm sure this will be in the next version(s) of rust{,ic}-mode, but this
fixes it for now.